Move some impls into a new module
Move `impl<T> Value for Option<T>` into its own file.
